server/requirements.txt
Flask
flask-cors
python-dotenv
youtube-transcript-api
requests # For making HTTP requests to Ollama
# Add any other specific LLM client libraries here, e.g., openai
